+++ .gitignore
@@ -18,3 +18,4 @@ config/database.yml
 .ruby-gemset
 .ruby-version
 ngrok
+Gemfile.lock
