+++ Gemfile
@@ -10,39 +10,37 @@ gem 'requirable', '~> 0.0.3'
 gem 'usecasing', '~> 0.1.10'
 gem 'sinatra', '~> 1.4.6'
 gem 'backports', '~> 3.6.4'
-gem 'rake'
-gem 'pg'
-gem 'gems', require: 'gems'
-gem 'rack-parser', require: 'rack/parser'
-gem 'hashie'
-gem 'rack-protection'
-gem 'kaminari', require: 'kaminari/sinatra'
-gem 'padrino-helpers'
-gem 'active_model_serializers'
-gem 'httparty'
-gem 'paranoia'
+gem 'rake', '~> 10.4.2'
+gem 'pg', '~> 0.17.1'
+gem 'rack-parser', '~> 0.6.1', require: 'rack/parser'
+gem 'hashie', '~> 3.4.1'
+gem 'rack-protection', '~> 1.5.3'
+gem 'kaminari', '~> 0.16.3', require: 'kaminari/sinatra'
+gem 'padrino-helpers', '~> 0.12.5'
+gem 'active_model_serializers', '~> 0.9.3'
+gem 'httparty', '~> 0.13.5'
+gem 'paranoia', '~> 2.1.3'
+gem 'pusher', '0.14.5'
 
 gem 'capistrano-ext', '1.2.1'
 gem 'capistrano', '2.15.5'
-gem 'rvm-capistrano', require: false
 
 # auth
-gem 'omniauth'
-gem 'omniauth-slack'
+gem 'omniauth', '~> 1.2.2'
+gem 'omniauth-slack', '~> 2.0.0'
 
-gem 'tux'
+gem 'tux', '~> 0.3.0'
 
 group :development do
-  gem 'pry', require: 'pry'
+  gem 'pry', '~> 0.10.1', require: 'pry'
   gem 'better_errors', '~> 2.1.1'
-  gem 'binding_of_caller'
-  gem 'webmock'
+  gem 'binding_of_caller', '~> 0.7.2'
 end
 
 group :test do
-  gem 'pry', require: 'pry'
-  gem 'rspec'
-  gem 'factory_girl'
-  gem 'database_cleaner'
-  gem 'webmock'
+  gem 'pry', '~> 0.10.1', require: 'pry'
+  gem 'rspec', '~> 3.3.0'
+  gem 'factory_girl', '~> 4.5.0'
+  gem 'database_cleaner', '~> 1.4.1'
+  gem 'webmock', '~> 1.21.0'
 end
