+++ app/commiker/v0/use_cases/sprints/communicate_pusher/base.rb
@@ -1,0 +1,28 @@
+module Commiker
+  module V0
+    module UseCases
+      module Sprints
+        module CommunicatePusher
+
+          class Base < UseCaseBase
+
+            context_reader :sprint
+
+            def perform
+              if Configs['USE_PUSHER'] == true
+                if sprint
+                  Pusher['commiker_dev']
+                    .trigger('sprints.show', Serializers::Sprints::Show.new(ctx.sprint))
+                else
+                  puts '!!!No sprint given to communicate, somethings wrong?!!!'
+                end
+              end
+            end
+
+          end
+
+        end
+      end
+    end
+  end
+end
