+++ app/commiker/v0/use_cases/stories/add_story_interaction/base.rb
@@ -7,18 +7,9 @@ module Commiker
           class Base < UseCaseBase
 
             depends Stories::FindOne,
-                    ParseParams
-
-            context_reader :story,
-                           :story_interaction_attributes
-
-            def perform
-              story_interaction_attributes.merge!({
-                interacted_at: Time.now
-              })
-
-              story.story_interactions.create(story_interaction_attributes)
-            end
+                    ParseParams,
+                    Persist,
+                    Sprints::CommunicatePusher::Base
 
           end
