+++ app/web/views/shared/_head_scripts.erb
@@ -1,3 +1,7 @@
+<!-- angular-js -->
 <script src="//ajax.googleapis.com/ajax/libs/angularjs/1.3.5/angular.min.js"></script>
+<!-- jquery -->
 <script src="//ajax.googleapis.com/ajax/libs/jquery/1.10.2/jquery.min.js"></script>
 <script src="//ajax.googleapis.com/ajax/libs/jqueryui/1.10.3/jquery-ui.min.js"></script>
+<!-- pusher-js -->
+<script src="//js.pusher.com/2.2/pusher.min.js"></script>
