+++ config/initializers/pusher.rb
@@ -1,0 +1,3 @@
+Pusher.url = \
+  "http://#{Configs['PUSHER_KEY']}:#{Configs['PUSHER_SECRET']}@" \
+  "api.pusherapp.com/apps/#{Configs['PUSHER_APP_ID']}"
