+++ public/templates/sprints/show/list.html
@@ -7,7 +7,7 @@
         class="media"
         style="position: relative; padding: 10px 5px 10px 10px;"
         ng-class="{ 'alert-warning': (user.stories.length == 0) }"
-        ng-repeat="user in sprint.users">
+        ng-repeat="user in sprintsShowCtrl.sprint.users">
 
         <div href="#" class="pull-left" style="padding-right: 20px;">
           <img alt="64x64" class="media-object img-circle" style="width: 64px; height: 64px;" src="{{user.avatar_72_url}}">
@@ -54,7 +54,7 @@
         class="media"
         style="position: relative; padding: 10px 5px 10px 10px;"
         ng-class="{ 'alert-warning': (user.stories.length == 0) }"
-        ng-repeat="user in sprint.users">
+        ng-repeat="user in sprintsShowCtrl.sprint.users">
         -->
         <div class="media-body" style="width: 75%;">
           <div class="media-heading" style="text-align: left">
@@ -71,7 +71,7 @@
             <ul class="list-group default_text">
               <li
                 class="list-group-item st_working"
-                ng-init="delivered = (!sprint.active && story.completion_percentage == 100)"
+                ng-init="delivered = (!sprintsShowCtrl.sprint.active && story.completion_percentage == 100)"
                 ng-class="{ 'delivered': delivered, 'not-delivered': delivered }"
                 ng-repeat="story in user.stories"
                 style="padding-bottom: 30px">
@@ -82,7 +82,7 @@
                     <span style="font-weight: bold; color: #4C91E4">[{{story.completion_percentage}}%]</span>
                   </span>
                   <span
-                  ng-if="!sprint.active"
+                  ng-if="!sprintsShowCtrl.sprint.active"
                   class="label label-{{delivered ? 'success' : 'danger'}}"
                   >{{delivered ? 'delivered' : 'not delivered'}}</span>
                 </div>
